pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Taras Nytiaha", email = "taras.nytiaha@gmail.com"},
]
dependencies = [
    "telethon>=1.28.5",
    "asyncio>=3.4.3",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"
license = {text = "MIT"}
